requirements.txt
streamlit==1.31.1
google-generativeai==0.8.3
python-dotenv==1.0.1
pandas==2.2.0
